requirements.txt
streamlit>=1.36.0
openai>=1.0.0
chromadb>=0.4.0
numpy>=1.24.0
scikit-learn>=1.3.0
plotly>=5.18.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
tiktoken>=0.5.0
streamlit-flow-component>=0.6.0
pypdf>=4.0.0
python-docx>=1.0.0
pandas>=2.1.0
gensim>=4.3.0
sentence-transformers>=2.2.0
torch>=2.0.0
pillow>=10.0.0
